Use one object with different ranges, change to uint everywhere.

+++ instance.go
@@ -86,7 +86,7 @@ func getInstanceId() string {
 }
 
 // generates an S3 key from the sha hash of the hostname, thread index, and object size
-func generateS3Key(host string, threadIndex int, payloadSize uint64) string {
+func generateS3Key(host string, threadIndex int, payloadSize usize) string {
 	keyHash := sha1.Sum([]byte(fmt.Sprintf("%s-%03d-%012d", host, threadIndex, payloadSize)))
 	key := fmt.Sprintf("%x", keyHash)
 	return key
